ARM: remoteproc: error handler api
Add the possibility to report an error to the generic
remoteproc framework in order to trigger the recovery
mechanism.

Currently the errors that can be reported are:
- RPROC_IOMMU_FAULT
- RPROC_WATCHDOG
- RPROC_BUS_FAULT
- RPROC_DIV_BY_ZERO
- RPROC_STACK_OVERFLOW
- RPROC_MEM_CORRUPTION
- RPROC_GENERAL

Each architecture will implement their way to receive
wichever error they can detect and report it to the
generic layer.

If the architecture is capable of detecting errors it will
register the callback rproc_error_handler_t in their reg_err_handler
implementation.

TODO: the recovery mechanism is pending and will be introduced
in subsequent patches.

Signed-off-by: Miguel Vadillo <vadillo@ti.com>